src/app/requirements.txt
databricks-sql-connector
plotly
